+++ crates/goose/tests/acp_server_test.rs
@@ -216,7 +216,6 @@ fn test_load_session_error_session_not_found() {
 }
 
 #[test]
-#[ignore = "TODO(lifei): decide Goose ACP loadSession mcpServers ownership semantics"]
 fn test_load_session_mcp() {
     run_test(async { run_load_session_mcp::<AcpServerConnection>().await });
 }
